Support custom theme on team sign-in page
