+++ app/components/Theme.tsx
@@ -1,41 +1,19 @@
 import { observer } from "mobx-react";
 import * as React from "react";
 import { ThemeProvider } from "styled-components";
-import { breakpoints } from "@shared/styles";
 import GlobalStyles from "@shared/styles/globals";
-import {
-  buildDarkTheme,
-  buildLightTheme,
-  buildPitchBlackTheme,
-} from "@shared/styles/theme";
-import { UserPreference } from "@shared/types";
-import useMediaQuery from "~/hooks/useMediaQuery";
+import { TeamPreference, UserPreference } from "@shared/types";
+import useBuildTheme from "~/hooks/useBuildTheme";
 import useStores from "~/hooks/useStores";
 import { TooltipStyles } from "./Tooltip";
 
 const Theme: React.FC = ({ children }) => {
   const { auth, ui } = useStores();
-  const isMobile = useMediaQuery(`(max-width: ${breakpoints.tablet}px)`);
-  const isPrinting = useMediaQuery("print");
-
-  const theme = React.useMemo(() => {
-    const customTheme = auth.team?.preferences?.customTheme || {};
-
-    return isPrinting
-      ? buildLightTheme(customTheme)
-      : isMobile
-      ? ui.resolvedTheme === "dark"
-        ? buildPitchBlackTheme(customTheme)
-        : buildLightTheme(customTheme)
-      : ui.resolvedTheme === "dark"
-      ? buildDarkTheme(customTheme)
-      : buildLightTheme(customTheme);
-  }, [
-    auth.team?.preferences?.customTheme,
-    isMobile,
-    isPrinting,
-    ui.resolvedTheme,
-  ]);
+  const theme = useBuildTheme(
+    auth.team?.getPreference(TeamPreference.CustomTheme) ||
+      auth.config?.customTheme ||
+      undefined
+  );
 
   React.useEffect(() => {
     window.dispatchEvent(
